nvme: implement namespace type support
Implement basic support for namespace types.

Disclaimer: Some bits might be missing as (i) the TP is not complete and
(ii) I did not go with a fine comb on it. The current implementation
allows to identify and allocate different namespace types on a sungle
device. Note that some TODOs are there to follow the normal kernel NVMe
bringup.

Signed-off-by: Javier González <javier.gonz@samsung.com>