blkzone: add nr_zones to zone properties
Allow a zone device to report the number of zones as part of the zone
properties. This allows user space to get the real number of zones thtough an
ICTL without having to do the calculations based on the device capacity, as this
becomes complicated when the zone sizes and capacities differ.

Signed-off-by: Javier González <javier.gonz@samsung.com>