nvme: add whitelisting infrastructure
If CAP_SYS_ADMIN is present, nothing else is checked, as before this
patch. if CAP_SYS_ADMIN is not present, take the decision based on
- type of nvme command (io or admin)
- nature of nvme-command (write or read)
- mode with which file was opened (read-only, read-write etc.)

io-commands that write/read are allowed only if matching file mode is
present.
for admin-commands, few read-only admin command are allowed and that too
when mode matches.

Signed-off-by: Kanchan Joshi <joshi.k@samsung.com>